RGW/standalone: Build without RadosStore
