+++ src/rgw/rgw_object_expirer.cc
@@ -31,7 +31,7 @@
 #include "rgw_formats.h"
 #include "rgw_usage.h"
 #include "rgw_object_expirer_core.h"
-#include "driver/rados/rgw_zone.h"
+#include "rgw_zone.h"
 #include "rgw_sal_config.h"
 
 #define dout_subsys ceph_subsys_rgw
@@ -88,8 +88,7 @@ int main(const int argc, const char **argv)
   ceph::async::io_context_pool context_pool{cct->_conf->rgw_thread_pool_size};
 
   const DoutPrefix dp(cct.get(), dout_subsys, "rgw object expirer: ");
-  DriverManager::Config cfg;
-  cfg.store_name = "rados";
+  DriverManager::Config cfg = DriverManager::get_config(false, g_ceph_context);
   cfg.filter_name = "none";
   std::unique_ptr<rgw::sal::ConfigStore> cfgstore;
   auto config_store_type = g_conf().get_val<std::string>("rgw_config_store");
